+++ packages/InlineEditors/src/Table/index.js
@@ -1,1 +1,2 @@
-export { default } from "./Table";
+export { status } from "../types";
+export { default, useTable } from "./Table";