+++ packages/InlineEditors/stories/Editing.stories.js
@@ -1,7 +1,16 @@
 import React from "react";
 import { getStoryName } from "storybook/storyTree";
 import Table from "./examples/Table";
-import { ListBoxSingleStory } from "./examples/ListBox";
+import TableErrorStory from "./examples/Table.error";
+import {
+  ListBoxSingleStory,
+  ListBoxMultipleWithError,
+  ListBoxMultipleWithLoading,
+  ListBoxMultipleWithSuccess,
+} from "./examples/ListBox";
+
+import { WithScreener as WithScreenerStory } from "./screener/Editing.screener.stories";
+
 import Input from "./examples/Input";
 
 export default {
@@ -12,10 +21,27 @@ export const WithInATable = () => {
   return <Table />;
 };
 
+export const WithInATableError = () => {
+  return <TableErrorStory />;
+};
+
 export const ListBoxSingleWithInlineEditing = () => {
   return <ListBoxSingleStory />;
+};
+export const ListBoxSingleMultipleWithError = () => {
+  return <ListBoxMultipleWithError />;
+};
+export const ListBoxSingleMultipleWithLoading = () => {
+  return <ListBoxMultipleWithLoading />;
+};
+export const ListBoxSingleMultipleWithSuccess = () => {
+  return <ListBoxMultipleWithSuccess />;
 };
 
 export const WithInput = () => {
   return <Input />;
+};
+
+export const WithScreener = () => {
+  return <WithScreenerStory />;
 };