+++ packages/CollapsibleCard/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [5d0db59]
+  - @paprika/icon@1.1.0-next.0
+  - @paprika/button@1.0.9-next.0
+
 ## 2.0.1
 
 ### Patch Changes