+++ packages/DataField/CHANGELOG.md
@@ -1,5 +1,21 @@
 <!-- content -->
 
+## 1.0.13-next.1
+
+### Patch Changes
+
+- Updated dependencies [5d0db59]
+  - @paprika/icon@1.1.0-next.0
+  - @paprika/button@1.0.9-next.0
+  - @paprika/data-grid@1.0.10-next.1
+
+## 1.0.13-next.0
+
+### Patch Changes
+
+- Updated dependencies [00693b0]
+  - @paprika/data-grid@1.0.10-next.0
+
 ## 1.0.12
 
 ### Patch Changes