+++ packages/DataGrid/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.0.10-next.1
+
+### Patch Changes
+
+- Updated dependencies [5d0db59]
+  - @paprika/icon@1.1.0-next.0
+  - @paprika/button@1.0.9-next.0
+  - @paprika/checkbox@1.0.9-next.0
+
+## 1.0.10-next.0
+
+### Patch Changes
+
+- 00693b0: Prevent data grid from throwing error.
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).