+++ packages/DateInput/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.9-next.0
+
+### Patch Changes
+
+- Updated dependencies [5d0db59]
+  - @paprika/icon@1.1.0-next.0
+  - @paprika/input@3.0.1-next.0
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).