+++ packages/DatePicker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.13-next.0
+
+### Patch Changes
+
+- @paprika/calendar@1.0.10-next.0
+- @paprika/date-input@1.0.9-next.0
+
 ## 1.0.12
 
 ### Patch Changes