+++ packages/ListBoxBrowser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.18-next.0
+
+### Patch Changes
+
+- Updated dependencies [5d0db59]
+  - @paprika/icon@1.1.0-next.0
+  - @paprika/list-box@3.0.3-next.0
+
 ## 1.0.17
 
 ### Patch Changes